Tests for day 1 problems
I got day 1 slightly wrong on my first submission as I missed one of the test cases so had missed the counting when there were more than 2 numbers of the same digit in a row.

Putting in tests for the test cases they give you would have helped, so I'm trying that retrospectively for day 1.