Start with trying some tests for day 2
I've tried to put the test data into some tests that will tell me whether I'm on the right track. It's interesting as I've had to "guess" at the final structure in order to write the test.

That could end up being wrong, so I'll refactor the test. I'm starting to understand this testing business. One step at a time.

My suspicion is that in the end I'll need to turn the input of numbers into a list of numbers.